Create asn1crypto_tests package, along with supporting tooling
Adds the following tasks:

 - python run.py build
 - python run.py version {pep440_version}

Tests may now be executed a number of different ways and will
automatically ensure the local copy of asn1crypto is used, if run from
a Git working copy, or archive of a working copy.

Versioning scheme switched from SemVer to PEP 440 since that is what
the Python ecosystem tooling supports.