Add __getinitargs__ method to util.timezone
If a timezone object is copied, its __reduce__ method is called which in
turn calls __getinitargs__ to get arguments for __init__ on the new
object.
If that function is missing, __init_ is called with too few parameters.

Problem can be triggered by calling copy or untag on UTCTime or GeneralizedTime
or any other object where the native value includes a datetime

Fixes #167