Fix parsing of ObjectIdentifier's first octet.
X.660 §7.6 specifies that the first arc of an OID must be 0 to 2 and
the second arc below arcs 0 and 1 must be 0 to 39. Arc 2 is not
restricted in that way.

The algorithm for parsing ObjectIdentifiers did not work for
OID 2.X with X >= 40.

This change also enforces above restrictions when setting an
ObjectIdentifier.