Fix missing bounds checks in parser
_parse is missing a host of bounds checks.  This means the parser would
raise IndexError instead of ValueError was it supposed to. Second, this
library cannot rely on Python's built-in bounds-checking because, when
parsing subelements, it passes in the whole string with a shorter
data_len. All bounds checks need to be against the shorter value.

Additionally, the length_only mode inconsistently checked that the
returned content_end value was valid. Restructure the code so that the
check can be applied uniformly.

Finally, this fixes the wrong numbers being passed into the exception in
one of the bounds checks.